refactor(xapiri): extract updateLogsTab + renderLogsTab → tab_logs.go
Mechanical move per ADR 0014. No behaviour change.
